+++ src/main/java/org/yesworkflow/data/TemplateVariable.java
@@ -2,10 +2,10 @@ package org.yesworkflow.data;
 
 public class TemplateVariable {
 
-    public final int id;
+    public final Long id;
     public final String name;
     
-    public TemplateVariable(int id, String name) {
+    public TemplateVariable(Long id, String name) {
         this.id = id;
         this.name = name;
     }
